@@ -4,10 +4,10 @@ build-backend = "hatchling.build"
 
 [project]
 name = "moonbit-up"
-version = "0.1.0"
+version = "0.2.0"
 description = "MoonBit toolchain manager for ARM64 systems with QEMU emulation support"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.11"
 license = {text = "MIT"}
 authors = [
     {name = "Damian Reeves"}
@@ -16,7 +16,6 @@ dependencies = [
     "typer[all]>=0.9.0",
     "rich>=13.0.0",
     "requests>=2.31.0",
-    "tomli>=2.0.0; python_version < '3.11'",
 ]
 
 [project.scripts]
