@@ -1,3 +1,3 @@
 """MoonBit toolchain manager for ARM64 systems."""
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
